Add experimental 'separation' statistic
Added a statistic that simply uses the difference of means
divided by the maximum difference of means across all permutations
of the data. This results in a statistic between 0 and 1 (inclusive)
where 1 represents the optimal grouping of the values (in terms of
means separation) and 0 represents the worst. I thought I'd try
this out, and on the test data, it doesn't seem to perform well at
all, so I'm not moving forward with it, but I'll leave it here for
now.